feat: oracle deployment
