+++ blockchain/package.json
@@ -32,7 +32,9 @@
     "dev": "hardhat node & yarn deploy --network localhost",
     "stop": "npx kill-port 8545",
     "compile": "hardhat compile",
-    "deploy": "hardhat compile && hardhat deployTest --network localhost",
+    "deploy:voting": "hardhat compile && hardhat deployVoting --network localhost",
+    "deploy:oracle": "hardhat compile && hardhat deployOracle --network localhost",
+    "deploy": "hardhat compile && hardhat deploy --network localhost",
     "verify": "hardhat verify",
     "test": "hardhat test",
     "test:coverage": "hardhat coverage",
